@@ -1,6 +1,6 @@
 {
   "name": "apollo-connector-mongodb",
-  "version": "2.0.2",
+  "version": "5.0.0",
   "description": "Apollo connector for mongodb",
   "main": "./dist/index.js",
   "scripts": {
@@ -21,30 +21,28 @@
     "extends": "./eslint.js"
   },
   "dependencies": {
+    "dataloader": "^2.2.2",
+    "lru-cache": "^9.1.2",
+    "meteor-random": "0.0.3"
+  },
+  "devDependencies": {
     "@types/mocha": "^2.2.48",
     "@types/proxyquire": "^1.3.28",
     "@types/sinon": "^4.3.0",
-    "dataloader": "^1.4.0",
-    "import-to-commonjs": ">=0.2.2",
-    "meteor-random": "0.0.3",
-    "mongodb": "^3.0.4"
-  },
-  "devDependencies": {
-    "@types/lru-cache": "^4.1.0",
     "@types/mongodb": "^3.0.7",
     "@types/power-assert": "^1.4.29",
     "assert": "^1.4.1",
     "coveralls": "^3.0.0",
-    "lru-cache": "^4.1.1",
     "mocha": "^5.0.2",
     "mocha-lcov-reporter": "^1.3.0",
+    "mongodb": "^5.0.0",
     "nyc": "^11.4.1",
     "power-assert": "^1.4.4",
     "proxyquire": "^2.0.0",
     "sinon": "^4.4.2",
     "sinon-test": "2.1.3",
     "ts-node": "^5.0.1",
-    "typescript": "^2.7.2"
+    "typescript": "^5.1.3"
   },
   "nyc": {
     "include": [
